Semi-radical change: removing all MAKES-specific objects leaving only OpenAlex

+++ EPPIReviewer4.2015/Server/BusinessClasses/MagRelatedPapersRun.cs
@@ -542,7 +542,7 @@ namespace BusinessLibrary.BusinessClasses
                     {
                         command.CommandType = System.Data.CommandType.StoredProcedure;
                         command.Parameters.Add(new SqlParameter("@MAG_RELATED_RUN_ID", ReadProperty(MagRelatedRunIdProperty)));
-                        command.Parameters.Add(new SqlParameter("@N_PAPERS", 0));
+                        command.Parameters.Add(new SqlParameter("@N_PAPERS", (int)0));
                         command.ExecuteNonQuery();
                     }
                     connection.Close();
